+++ src/profile/ProfileMetadataSchema.ts
@@ -96,7 +96,7 @@ export type ProfileMetadata = {
   /**
    * A cryptographic signature of the `lens` data.
    *
-   * @experimental DO NOT use ye
+   * @experimental DO NOT use yet
    */
   signature?: Signature;
 };