+++ src/publication/3D.ts
@@ -121,7 +121,7 @@ export type ThreeDMetadata = MarketplaceMetadata & {
   /**
    * A cryptographic signature of the `lens` data.
    *
-   * @experimental DO NOT use ye
+   * @experimental DO NOT use yet
    */
   signature?: Signature;
 };