+++ src/publication/ArticleSchema.ts
@@ -65,7 +65,7 @@ export type ArticleMetadata = MarketplaceMetadata & {
   /**
    * A cryptographic signature of the `lens` data.
    *
-   * @experimental DO NOT use ye
+   * @experimental DO NOT use yet
    */
   signature?: Signature;
 };