+++ src/publication/LinkSchema.ts
@@ -65,7 +65,9 @@ export type LinkMetadata = MarketplaceMetadata & {
    */
   lens: LinkMetadataDetails;
   /**
-   * The signature.
+   * A cryptographic signature of the `lens` data.
+   *
+   * @experimental DO NOT use yet
    */
   signature?: Signature;
 };