+++ src/publication/SpaceSchema.ts
@@ -84,7 +84,7 @@ export type SpaceMetadata = MarketplaceMetadata & {
   /**
    * A cryptographic signature of the `lens` data.
    *
-   * @experimental DO NOT use ye
+   * @experimental DO NOT use yet
    */
   signature?: Signature;
 };