+++ src/publication/TransactionSchema.ts
@@ -93,7 +93,7 @@ export type TransactionMetadata = MarketplaceMetadata & {
   /**
    * A cryptographic signature of the `lens` data.
    *
-   * @experimental DO NOT use ye
+   * @experimental DO NOT use yet
    */
   signature?: Signature;
 };