+++ src/publication/VideoSchema.ts
@@ -79,7 +79,7 @@ export type VideoMetadata = MarketplaceMetadata & {
   /**
    * A cryptographic signature of the `lens` data.
    *
-   * @experimental DO NOT use ye
+   * @experimental DO NOT use yet
    */
   signature?: Signature;
 };